crawler/requirements.txt
fastapi
uvicorn
playwright
groq
httpx
requests
beautifulsoup4
python-dotenv
psutil
websockets
browser-use
langchain-groq
langchain-openai
langchain-huggingface
langchain-google-genai
tenacity
